refactor(client): share structured command collection
