+++ src/client/transports/MemoryTransport.ts
@@ -2,7 +2,7 @@ import * as Cli from '../../Cli.js'
 import { createClientDiscover } from '../../internal/client-discover.js'
 import { createClientLocal } from '../../internal/client-local.js'
 import { createClientRequest } from '../../internal/client-request.js'
-import * as RuntimeContext from '../../internal/client-runtime-context.js'
+import * as RuntimeContext from '../../internal/runtime-context.js'
 import { ClientError } from '../ClientError.js'
 import type * as Discover from '../Discover.js'
 import type * as Local from '../Local.js'
